Фикс сохранения настройки масштаба оверлея

+++ dev/settingsController/settingsController.cpp
@@ -22,7 +22,7 @@ void SettingsController::initializeSettings()
     m_settings->noFog = ss_stats_settings->value("game/no_fog", true).toBool();
     m_settings->smallGamePanelActive = ss_stats_settings->value("client/small_pannel_active", false).toBool();
     m_settings->showGamePanelPreset = ss_stats_settings->value("client/show_game_panel_preset", true).toBool();
-    m_settings->scale = ss_stats_settings->value("client/scale", 1.0).toFloat();
+    m_settings->scale = ss_stats_settings->value("client/scale", 1.0).toDouble();
     m_settings->lastReadedNewsMessageID = ss_stats_settings->value("client/last_readed_news_message_id", "").toString();
     m_settings->lastReadedEventsMessageID = ss_stats_settings->value("client/last_readed_events_message_id", "").toString();
     m_settings->russianFontsInstalled = ss_stats_settings->value("mods/russian_fonts_installed", false).toBool();
